move pjsuafilehandler
